backend/fastapi/requirements.txt
fastapi>=0.109.2
uvicorn[standard]>=0.25.0
sqlalchemy>=2.0.23
pydantic>=2.5.3
python-dotenv>=1.0.0
openai>=1.10.0
qdrant-client>=1.7.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4